requirements.txt
# Core Dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0

# MCP Server
mcp>=0.9.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.17.0

# Database (sqlite3 is built-in to Python, no installation needed)

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
